requirements.txt
aes-package==0.2
alembic==1.14.1
aniso8601==7.0.0
anyio==4.8.0
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
pluggy==1.5.0
promise==2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytz==2025.1
redis==5.2.1
Rx==1.6.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
